+++ banzai/tests/test_dark_maker.py
@@ -2,7 +2,7 @@ import numpy as np
 
 from banzai.utils import stats
 from banzai.dark import DarkMaker
-from banzai.tests.utils import FakeContext, throws_inhomogeneous_set_exception
+from banzai.tests.utils import FakeContext, handles_inhomogeneous_set
 from banzai.tests.dark_utils import FakeDarkImage
 
 
@@ -27,20 +27,16 @@ def test_header_cal_type_dark():
     assert images[0].header['OBSTYPE'].upper() == 'DARK'
 
 
-def test_raises_an_exception_if_ccdsums_are_different():
-    throws_inhomogeneous_set_exception(DarkMaker, FakeContext(), 'ccdsum', '1 1', calibration_maker=True)
+def test_returns_null_if_ccdsums_are_different():
+    handles_inhomogeneous_set(DarkMaker, FakeContext(), 'ccdsum', '1 1', calibration_maker=True)
 
 
-def test_raises_an_exception_if_epochs_are_different():
-    throws_inhomogeneous_set_exception(DarkMaker, FakeContext(), 'epoch', '20160102', calibration_maker=True)
+def test_returns_null_if_nx_are_different():
+    handles_inhomogeneous_set(DarkMaker, FakeContext(), 'nx', 105, calibration_maker=True)
 
 
-def test_raises_an_exception_if_nx_are_different():
-    throws_inhomogeneous_set_exception(DarkMaker, FakeContext(), 'nx', 105, calibration_maker=True)
-
-
-def test_raises_an_exception_if_ny_are_different():
-    throws_inhomogeneous_set_exception(DarkMaker, FakeContext(), 'ny', 107, calibration_maker=True)
+def test_returns_null_if_ny_are_different():
+    handles_inhomogeneous_set(DarkMaker, FakeContext(), 'ny', 107, calibration_maker=True)
 
 
 def test_makes_a_sensible_master_dark():